requirements.txt
numpy==1.20.1
Pillow==8.1.0
scipy==1.6.1
matplotlib
prox-tv==3.2.1
torch==1.8.0
